Use a bit simpler jQuery to check how many images have been posted.
Wait 3 seconds instead of 30 seconds before setting canPost to true
again. It's annoying to try to upload an image, get an "image has
already been posted" or "file too large" error, post a different image 5
seconds later, and have it not sync your name. It's better to sync the
name more often with filenames that don't get posted than it is to miss
syncing your name with filenames that do get posted. Proper solution
would be better integration with 4chan X or just listen to the Ajax
calls 4chan X makes, but I'm not sure how to do that easily and in a
cross-browser supporting way yet.

Treat usedFilenames as a set, which is a bit more efficient for querying
membership.